Copy old release pipeline to 4.10.3
See https://github.com/gradle/gradle-private/issues/1698

We released 4.10.2 on Sep 17, after that we changed a lot of stuff.
Now we want to release 4.10.3 so we copy the whole `Gradle_Check` directory
on Sep 16 https://github.com/gradle/gradle/tree/b88019605bf then change it
to `Gradle4_Check` to perform the release process.