Tag CACHE_MISS upon unexpected cache miss (#7869)
This is the first-step implementation of https://github.com/gradle/gradle-private/issues/1681

Previously sometimes we see unexpected cache misses in the build, which we want to monitor. This PR marks all non-skipped `AbstractCompile` and `ClasspathManifest` tasks as `CACHE_MISS` if current build is not `compileAll`, so that this build can be searched in build scan later.