Remove build-receipt.properties from Wrapper jar
Prior to this commit, the `gradle-wrapper.jar` was created when
executing the `wrapper` task by copying the `gradle-wrapper.jar`
classpath resource and adding a Wrapper-specific
`build-receipt.properties` to it. The properties file only contained a
`versionNumber` property that was solely used to build the User-Agent
for downloading the distribution.

Now, the User-Agent no longer includes the version number. Thus, the
Wrapper-specific `build-receipt.properties` file is no longer needed.

This change will make it easier to publish the Wrapper jar or, as a
first step, its checksum alongside the distributions.

Resolves #6632.