Remove WorkIdentity
Instead of assigning arbitrary ids to all work items that are part of
the execution graph, include the `Nodes` in
`ExecuteTaskBuildOperationDetails` and
`ExecuteScheduledTransformationStepBuildOperationDetails` so they can
be queries for dependency successors directly. This eliminates the need
to query `TaskExecutionGraphInternal` or `ExecutionPlan` for the node
in order to determine its dependencies.