Add sign to performance test graphs (#7870)
This PR did some improvements for https://github.com/gradle/gradle-private/issues/1685 . Previously we display confidence and difference in different graphs, which make it hard to recognize whether a high confidence result is improvement or regression. This PR adds a background color for the difference graph and sign to the confidence graph.