Improve content of Maven migration guide
I have introduced a _General guidelines_ section that's modeled on the one in
the Ant migration guide. This merges and replaces a couple of the existing
early sections.

I have also added a section that tries to explain the difference between Maven
and Gradle when it comes to the build lifecycle, pointing readers at lifecycle
tasks as a comparative feature to the Maven lifecycle phases.

Finally, I cleaned up several sections -- filtering resources, configuring
integration tests -- and copied the _Further reading_ section from the Ant
migration guide.