Improve application plugin docs
In 45dcc84eb8d8fcf1c2fa94a9b8e2b54907d1b39b, an extension class for the Java
application plugin, JavaApplication, was created with intent to deprecate the
existing convention properties for said plugin.

This commit takes us closer to that by describing the convention properties as
deprecated in the application plugin userguide.

1. Create DSL docs for JavaApplication extension class.  This fields of this
class are currently identical to the convention properties.
2. Add missing field to both convention and extension DSL docs.
3. Give a brief example in JavaApplication javadoc illustrating that the
extension is applied to the project under the name 'application'.
4. Add missing links to both the convention and extension DSL docs in the
application plugin userguide.  Describe the convention properties as deprecated
in favor of the extension.

Closes #7939.

Signed-off-by: Ian Kerins <ianskerins@gmail.com>