Do not disable local cache on CI
We used to disable the local cache on CI for two reasons:

- without a local cache the remote backend learns about every cache reuse, and can prevent evicting an entry that is actually used everywhere,
- storing local results in the cache can lead to different artifacts used for the same cache key on different machines.

These pose some small risk of introducign cache misses, but the downside of not using the local cache is that we download the same artifacts to the same build agent many-many times. Let's see if reintroducing the local cache on CI breaks anything.