Make Gradle's own build reproducible
When running two builds of the same sources of Gradle,
outputs were not binary equivalent, this was primarily
due to zip files not being equivalent due to both
file ordering and timestamps in the file metadata.

This changes makes it so that running two gradle builds
with the same buildTimestamp
(e.g. -PbuildTimestamp="20181220000000-0800") produce the same output.

Having a reproducible build will make it easier to verify that
the official distributions match the sources on github.

Following changes were needed:

* switch to the kotlin-dsl provided zipTo methods instead of Ant's
  zip helper for the tasks that package the kotlin-dsl dependencies as zipTo
  correctly ignores file metadata
* set all AbstractArchiveTasks preserveFileTimestamp=false
  & reproducibleFileOrder=true to ensure jars created do not have file ordering
  or metadata changes from build to build

Signed-off-by: Jeffrey van Gogh <jvg@google.com>