Ensure completed() and failure() does not error for skipped tests
When a test class completes, all the in-progress tests have their
`TestState` set to `SKIPPED`. In some cases, there can be a race
condition between an in-progress test getting marked as `SKIPPED` and
the test actually completing/failing. This can happen, for example,
due to the test class having a timeout class rule.

The solution is for `StateTrackingTestResultProcessor` to keep track
of which tests have been skipped, and short-circuit if these tests
try to `completed()` or `failure()` later.

Issue: #8112

Signed-off-by: Leonid Gaiazov <leonid.gaiazov@appian.com>