Disallow '-J' flags in `CompilerOptions.compilerArgs`
`javac` already fails if -J flags (i.e. those affecting the JVM that is running
the compiler) appear in command-line argument files as opposed to directly on
the command line, but does so with a particularly useless error message.

This validation thus mainly serves to improve the error messaging, offering the
right alternative API for supplying such flags.

Signed-off-by: Ian Kerins <ianskerins@gmail.com>