Add a DSL for registering features on Java libraries
This commit makes it easier for a user to register optional features
in the Java ecosystem. Usually, in this world, optional dependencies
are used because for some reason it is not possible to split a project
into subprojects, and there are dependencies which are only required
at runtime if some feature is actually used by the consumer.

We make it easier to support this scenario by automatically setting
up a set of configurations which are used to support optional
features. Features, like Java libraries, may expose both an API
and an implementation. Consumers will select the right variants just
like for libraries.

The API makes it possible to support more scenarios, for example
if the feature is not implemented in the main source set. However,
this is not yet implemented. This means that the only scenario this
commit supports is the case where the main and optional features
are developed in the same, main, source set. Therefore, the artifact
for an optional feature will be the _same_ as the main one.