Add support for features using their own source set
This commit makes it possible to register a feature variant
which uses its own source set. In this case, the source set
needs to be created first. Then when registering the feature,
Gradle will make sure that a Jar exists corresponding to the
source set. If it doesn't, a task with the appropriate name
is created and attached to the outgoing configurations.