Make sure constraint attributes are merged with edge attributes
This commit fixes #8306 by making sure that if we resolve a constraint
edge and need to select a configuration based on attributes, we
consider the attributes of at least one real edge on it before resolving.

This allows the engine to consider attributes on constraints as requirements,
but also not to forget about edge constraints if they were any.