Rebaseline (non)-abi change performance tests
The (non-) abi change performance tests on the large monolithic java
project seem to be exposing a flakiness on `master`.
I rebaselined to see if that helps to mitigate the flakyness.
I checked the tests for the large multi project, and they don't seem
to have a regression, that's why I rebaseline both tests.