Add test output only once to cache artifact
Both reports have been added to the cached artifact by accident, since
the reports are declared as nested parameters via `enabledReports` on
the `ReportContainer` and the getters on `TestTaskReports`.

The annotated getters on `TestTaskReports` also prevented the test task
to be up-to-date when only the configuration for disabled reports
changed.