Fix unit test for UNIX start script generator
== DETAILS

The unit test "unix script uses unix line separator" had a
hardcoded value for how many lines it expected the generated
script to be.

Since my last commit deleted several lines, the test broke.

I've updated the test logic so it is no longer tied to the length
of the unix script.

Signed-off-by: Nathan Strong <nstrong@tripwire.com>