Replace :patchExternalModules with :kotlinCompilerEmbeddable
Only the kotlin-compiler-embeddable jar needs to be patched now.
Some patch actions were not needed anymore.

There's now a :kotlinCompilerEmbeddable subproject responsible for
building the patched kotlin-compiler-embeddable jar.
Only :kotlinDsl depends on it.

This greatly simplified the setup in the root project build script.
Various places required to be adjusted to account for the new setup.

The patching itself was rewritten to work in a streaming fashiong to
reduce i/o operations and disk usage. It was also moved up the
dependency tree.

All in all this should provide with a noticeable build speed up.

Signed-off-by: Paul Merlin <paul@gradle.com>