Deprecate `JavaLibrary` in favor of public API
This commit makes use of the newly introduced software component
factory to support publishing of Java components using a public
API only.

There are a few consequences to this change, mostly around Gradle
metadata:

- variants `api` and `runtime` are now named `apiElements` and
`runtimeElements`
- it is no longer possible to warn the user when publishing a
variant without capability. This might change later if we decide
to _always_ publish capabilities (and optimize on read)