Implement Gradle metadata marker in published pom/ivy files
This commit implements a performance optimization for Gradle metadata.
Given that today there's no library published in any repository with
Gradle metadata, it's much more likely to find a POM (or Ivy) metadata
file for an external dependency, rather than a Gradle metadata file.

If we decided to add `gradleMetadata()` sources by default to all
repositories, then we would probably introduce a performance regression
to a lot of builds, because we would first try to get Gradle metadata,
then fail, and search for POM/Ivy files.

To avoid this, whenever a library is going to be published with Gradle
metadata, we will introduce a _marker_ in the published POM or Ivy
file. When Gradle _resolves_ such a dependency, it will parse the POM
file and look for the marker. If it finds it, then it will _redirect_
to use Gradle metadata instead. This means that when Gradle metadata is
present, we will pay the price of looking for a POM or Ivy file first,
start parsing, only to discover we should parse Gradle metadata. This
should be ok in the beginning, knowing that if `gradleMetadata()` is
added, then we would systematically look at Gradle metadata first.

This means that this is a _temporary_ solution, until Gradle metadata
becomes widespread. So "temporary" should be understood as several
months, if not years.

The marker introduced in POM and Ivy files is _neutral_ for both Ivy
and Maven. By this we mean that it uses an XML comment. While not super
nice, we couldn't use a custom namespace because both Ivy and Maven
fail when parsing this. Properties were considered too, but Ivy doesn't
support them so for consistency both models use comments.

It's worth noting that we will still _completely parse_ the POM or Ivy
descriptor. It's a waste of time, but it helps in case we find a marker
but that for some reason the Gradle metadata file is absent. In this
case we fallback on the model we found.

This change also introduces a change in the semantics of the incubating
metadata sources API: those should be considered _hints_, and not strong
statements anymore.

Finally, should a producer want to avoid publishing Gradle metadata,
it's now possible to disable the task that creates the metadata file.