Fix disambiguation of platform/library
This commit fixes the fact that the engine will not let us
choose both the library and platform modules of a derived
POM component, because they don't have the same capabilities.
The fact they don't have the same capabilities is precisely
to allow them both to be present. However, as soon as we
have this, the `platform` keyword is not sufficient to
express the fact we allow both the implicit capability and
the implicit "derived" capability of a platform. For this
purpose we introduced the `ShadowedCapability` interface
which is _only_ used in this context.