Fix and re-enable :kotlinCompilerEmbeddable:clean
The CheckKotlinCompilerEmbeddableDependencies task had a property
getter calling DefaultTask.getTemporaryDir() which has the side effect
of creating that directory on each call. It happened very often that
this getter was called concurrently to the clean task, by overlapping
outputs detection maybe, causing the clean task to fail with "unable to
delete file".

This commit changes that task property to a simple field with no side
effects, moving the directory creation to the task action.

Signed-off-by: Paul Merlin <paul@gradle.com>