Fix equals/hashcode of mutable classloaders
When a parent was added to a MultiParentClassLoader, its hashcode
would change, meaning that it could no longer be found in any Sets
or Maps we put it in. This lead to severe memory leaks when used in
conjunction with configure-on-demand, where adding parents late is common.

These classloaders now use an identity hashcode and equals implementation.