Fix apiElements declaration
It was extending the runtime configuration, but instead needs to extend
the compile one.
This was causing dependencies declared in the runtime configuration to
appear as compile in produced metadata.
This change is however a breaking one as it will impact projects that
relied on dependencies declared in the runtime configuration to be
available for compiling dependent projects.