Fix version mapping backward compatiblity
This commit reworks how the default configurations are used
whenever mapping to resolved versions. Because of the
introduction of `JAVA_API_JARS`, any previous code which was
using `JAVA_API` in a version mapping block wouldn't match
anymore, because the outgoing variants are now using the
attributes from the configuration instead of the ones hard-coded.

So now we're using the attributes schema to find the most
appropriate mapping strategy, and use a dedicated path to set
the default value.