Add better change reporting for classpath inputs
Gradle should never return "modified" for jars which have different
absolute path and different hashes, since this makes it very hard
for incremental tasks to consume the changes.

We try to provide good changes for removed/added jars on the classpath
and modified jars with the same path as long as the order does not
change.

This addresses most of #1931.