Improve attribute disambiguation rule
This commit implements a smarter strategy for attribute disambiguation,
whenever _some_ variants have missing information. By missing information,
we mean that some variants carry at least one extra attribute, and others
do not. Because we consider the absence of an attribute as a compatible
match, BUT that when we disambiguate we only consider the variants which
actually provided a value, there was no way to prefer a match which
is "closer" to the request.

See #6747