Deduplication of serialized metadata
When writing realised Maven metadata, when a cacheable
ComponentMetadataRule is used, a lot of duplication happened around
MavenDependencyDescriptor.
This was caused by the amount of derived variants computed in for Maven
dependencies, which effectively have a very common set of dependencies.

This commit changes that for derived variants and configurations by
de-duplicating MavenDependencyDescriptor instances for serialization.

Issue #8311