Fix `enforcedPlatform` having dependencies instead of constraints
This commit fixes a major issue with `enforcedPlatform`: it was
using _dependencies_ instead of constraints. The consequence was
that all dependencies defined in a POM file would be visible in
an `enforcedPlatform`, when they should never have been there.

It also makes sure that derived plaforms do NOT carry any artifact.

Fixes #8548