Introduce `TargetJavaPlatform` attribute
This commit adds a new Java ecosystem attribute, corresponding
to the minimal target platform a producer supports. It's often
the case that Java libraries are published with different
classifiers for different target platforms. However, a classifier
is not enough. With this attribute, we now have a proper model
of the producer and consumer side.

By default, the producer will not ask for a specific target
platform, but will only accept those which are lower than or
equal to the current Gradle runtime version. However, if the
consumer says something, it will then select the most appropriate
variant based on the producer attributes.

In the Java world, a consumer can use Java libraries produced
for older versions of Java, but not newer versions. This rule
is baked in as the default compatibility rule. Disambiguation
will then chose the closest version.