Add an opt-out to automatic target jvm setting
Builds like Gradle have "incorrect" module setups: some modules
built for Java 6 have dependencies on Java 8 modules. This, in
general, shouldn't be allowed. However in Gradle it's "fine"
because we know that when we run on Java 6, the classes which
actually need Java 8 features or dependencies are not in use.