Proper propagation of flag when modifying metadata
The `ConfigurationBoundExternalDependencyMetadata` was not properly
propagating the `alwaysUseAttributeMatching` state when the metadata was
modified.
This caused legacy Maven / Ivy interop to kick in, as happened before Maven
was fully moved to variant aware dependency management. One test needed
to be modified, which in reality should have been changed at the 5.0
release.

Fixes #8586