Implement classifier to variant mapping
This commit implements a strategy to disambiguate
variants whenever the consumer uses a classifier
in the dependency descriptor. This can be the case
if a Maven library depends on a Gradle library published
with Gradle module metadata, or even if a Gradle
library depends on another Gradle library but makes
use of a classifier (when it should really be using an
attribute).