Check fewer times if dependencies complete
If a unit of work finishes, only anything which depends on that unit
of work can become ready for execution.

There is the special case of tasks in other builds which do not notify
the current task execution plan when they finish executing. So we need
to keep polling for those.