Change path sensitivity for Java compile to relative
The Java compiler care only about the name of the java file, not
necessarily about the relative path. It relies on the relative path
to find other classes in the same package.

There is also a problem for our incremental Java compiler, which does
not remove stale files when a file is moved in the hierarchy (#8874)

Therefore, the path sensitivity is changed to RELATIVE to avoid those
problems.