Ignore Kotlin compiler warnings in debug logs
When compiling the Gradle API, the deprecation warnings cause Kotlin
compiler warnings in the info log. When a test runs with debug logs,
those compiler warnings are treated as Gradle deprecation warnings
and make the test fail. Now the regular expression for matching
Kotlin compiler warnings has been changed, so it ignores the warnings
on the debug/info log as well.