Fix selection of target configuration with Ivy+constraints
This commit fixes an edge case when a dependency doesn't
use variant awareness, but rather an explicit target
configuration. In particular in the case of Ivy, the fact
of adding a constraint made the build fail with an error
stating that no matching configuration could be selected,
with the list of configurations and their attributes, but
no required attribute.

This case happened when the target module didn't provide
a "default" configuration.

Now we avoid selecting a configuration for constraint edges,
unless they are well known cases (virtual platforms). This
should have a positive impact on performance too.