Don't use static state
A compiler daemon is created per worker process and there isn't a
way for a compiler to execute 2 builds concurrently, or compile
two projects of a single build concurrently. This means that if
the scala compiler is reused, we can just check a regular field
and decide to reuse the cache or not.