Support exit on exception in `BlockingHandler.waitForAllPendingCalls()`
This allows a test that uses `waitForAllPendingCalls()` to pass a
failure tracker to the method. The tracker can be used to capture an
unexpected early exception. Instead of waiting for something that
might never happen, because of the unexpected exception,
`waitForAllPendingCalls()` will then terminate by throwing that
exception instead of eventually timing out.