Mark known flaky performance test in report (#8997)
This closes https://github.com/gradle/gradle-private/issues/1735

We want to mark the known performance test scenarios in the report even when they pass. This can give us insights of how the scenarios work previously. In this way, we can easily know, "this scenario fails, ah, it's marked as flaky", or, "this scenario is improved, but hold on, it's marked as flaky, let's not rebaseline it".