Add regression tests for clean dexing transform
We add a test for a clean build with dexing transforms.
That requires to clean up the transform caches between
runs.
This test shows us the improvements of running external
transforms in parallel.

We also add `phthalic:assembleDebug` as a task for testing
if there is only one long running task for one variant.