Handle cases when we have both 'is'-getters and 'get'-getters
Groovy generates both an 'is' and a 'get'-getter for boolean properties, and we shouldn't warn in this case. However, if the user defines both getter methods, we should require at least one of them to be explicitly ignored.

We need to figure out which method we need to call to get the value of the property. In the Groovy-generated case we can safely pick the 'get'-getter, because they both do exactly the same. In the user-specified case we need to be told which (if any) to use, as we have no means to make this decision on our own.