Use consistent ordering for module exclusions
This commit updates the `MergeSet` implementation to extend
from linked hashset instead of hashset, so that the order of
elements is consistent. In theory, it shouldn't matter, but
practice has shown that there must be a bug somewhere else
that is at least partially mitigated by using consistent
ordering here.

In practice what it means is that it is possible to get
inconsistent classpath ordering in multi-project builds
depending on the order of resolution of projects, when
those projects have lots of common exclusions in transitive
dependencies. The considtion to reach this bug are unclear
despite our efforts to find out, so this commit is really
just a workaround at this point.