Exclude workspace names when naming eclipse projects
Projects in eclipse workspaces need unique names. Gradle already ensures
this for projects in the same build. This change makes EclipseModelBuilder
a ParameterizedToolingModelBuilder and adds parameter objects for eclipse
workspace and project information. It furthermore adds a second
deduplication pass in EclipseModelBuilder that includes external eclipse
project names. This in turn reduces the amount of naming conflicts in
mixed eclipse workspaces.

Signed-off-by: Christian Fraenkel <fraenkelc@users.noreply.github.com>