De-duplicate dependencies before iterating
This fixes a bug whenever the same dependency is seen multiple
times on the same component: we're incorrectly tracking the number
of hard edges to this dependency. By deduping, we avoid the creation
of multiple edges to the same thing, which makes things faster and
fixes this bug.

Deduplication is done on `NodeState` because it's too expensive
to do it ahead of time, on configuration metadata, in particular
because of the lazy computation of dependencies and application
of rules.