Handle cycles in the "bean" object graphs written to the instant execution cache.
Also preserve object identity for the "beans" for a given task. That is, if 2 fields for a task reference the same instance of a "bean" when written to the cache, then they will also reference the same instance when read from the cache. Identity is intentionally not preserved across task instances, to keep the tasks isolated.