Remove check that's incompatible with IDEA 2019.2
IDEA 2019.2 uses the new PhasedAction API in order to
get a tooling model and run some tasks in the same step.
This API uses a projectsEvaluated listener under the hood,
which fires before the check we had in place to detect the
end of configuration time. As a result, we would falsely
fail, saying that IDEA was "resolving dependencies at config
time".