Fix inconsistent serialization state when resolution cache expires
Resolution results are cached, but the cache may expire or entries
in the cache may be evicted, in case there are many configurations
resolved. This has the consequence that the streaming resolution
builder root factory may be called multiple times for the same
configuration, and therefore the file may be read several times
within a build.

This commit fixes a bug with attribute serialization where the state
of the serializer would make the 2d read incorrect.