Fix incorrect reduction of intersections
Simplification of intersections in the normalizing exclude
factory is done by pairs. There was a bug in the reduction
algorithm, that would cause the reduction result to be
wrong because we used the wrong, non simplified, exclude spec
for merging whenever a merge occured, to reduce on the next
item in a list.