Do not use heuristic to tell if we should revisit a node
The previous implementation assumed that if we visited a node once,
and then revisit later, and that the exclusion filter "excludes
the same things if we ignore artifacts", then we don't need to revisit.

The problem with this approach is that it's a heuristic, that in
practice compares the exclude specs. But, two exclude specs can
be _equivalent_ (they would exclude the same things) but not _equal_
(their structure is different, because for example they were reduced
differently because of ordering).

This commit removes this faulty `equalsIgnoreArtifact` method in
favor of a direct comparison of the result of excluding edges. If
both specs are equivalent, they will exclude the same edges.