Record line numbers for each plugin request
When compiling, transform each `id(String)` method call into
`id(String, int)`, allowing the line number to be captured.
As the syntax of the plugins block is still completely restricted,
there is no risk of another `id` method being incorrectly transformed.

If/when we decide to relax the restrictions on syntax, we'll need to
do something smarter here.