Add requested attributes to the resolution result
This is done so that it is possible to find out what
attributes were requested/compatible/missing in variants
based on the consumer attributes, without having to
resort on finding the configuration being resolved.

This is mostly useful in build scans where the plugin
listens to every resolution result, but does not necessarily
know what configuration was resolved to generate this
result.