Deprecate declaring incremental tasks with no outputs
An incremental task should always declare an output. Gradle
automatically declared an upToDate spec for incremental tasks to make
sure this is always the case.

Actually, the implementor of the task should add the spec if the task
does not declare any outputs. We expect nearly all incremental tasks
to already declare outputs.