Fix compilation of Groovy projects depending on a library
This commit introduces a new "library elements" type corresponding
to the classes AND resources of a library. Groovy projects now tell
that they needs both the classes and resources (because the Groovy
compiler requires some descriptors).

Currently the implementation is basic as it will fallback to a jar
thanks to the attribute compatibility rules. However, it would be
better to provide a variant which actually packages only the classes
and resources as directories instead of within a jar.

Fixes #9872