Fix resolved versions of substituted dependencies
This commit fixes an inconsistency when publishing resolved
versions of a component. If that component happens to be
substituted, which would be strange for a first level dependency
but nevertheless possible, then we wouldn't find the target
module in the resolution result, and the outcome would be that
POM/IVY/Gradle Module metadata files would all have an empty
version for a substituted dependency.

With the change, we will now also look for dependencies in
the resolution result, and if one matches the original dependency
coordinates, then we use its resolved component as the result.

This allows us to _substitute_ the result with complete coordinates
in the metadata files.

Fixes nebula-plugins/gradle-nebula-integration#62