Check for already registered model in TaskContainer
A ModelMap<Task> is a shadow version of the TaskContainer inside the
software model. Although difficult, it is possible to create an element
in ModelMap<Task> before creating the same element in TaskContainer.
This creates a duplicate instance that, in some case, cause both
instance to be used in the task graph causing issues.