Add an outgoing variants report
This commit introduces an `outgoingVariants` task which generates
a report of the outgoing variants of a project. This is useful
for debugging and understanding variant aware resolution.

It becomes possible for a plugin author to understand what variants
are already produced by a project based on the plugins applied, and
what artifacts are attached to what variant.