Fail on unpack failure
Previously we used to try a complicated dance to continue even if unpacking a cached result failed. We did this because wanted to avoid a corrupted cache entry breaking a build. However, with real-world usage it looks like hiding these problems causes more grief, and adds quite some complexity to the process. So we are removing this mitigation now.