Improve rendering of artifacts in outgoing variants report
This commit changes how artifacts are rendered, so that if they
are associated with a type (either explicit or derived), we show
the corresponding `artifactType` attribute.

This helps understanding a bit what can happen in an Android build
where multiple variants may have the same attributes, but only
differ on the artifact type derived from the artifacts carried
by the variant.