Remove special repository default behavior with GRADLE_METADATA flag
The default behavior is now that we look for 'traditional' metadata
first, but find Gradle metadata if a marker is present.
We won't look for anything else if the 'traditional' metadata is not
found. This increases performance for the cases where one repository
is asked for many modules it does not contain.
Users can opt into different behavior using `gradleMetadata()` or
`artifact()` when defining their repositories.