Improve instant execution support for Java serialization
Properly handle deserialization of graphs containing a mix of
`java.io.Serializable` beans and plain beans by introducing
`ReadContext.immediateMode`. In immediate mode, reentrant `read` calls
are not suspended for compatibility with the non-suspending semantics
of `ObjectInputStream.readObject`.

The downside of non-suspending semantics is the risk of
`StackOverflowErrors` caused by deep graphs.