Ensure that no-isolation workers run with the classloader of the submitting thread
This fixes a problem where, in a composite build, a worker can get a context
classloader set to a classloader from another project, which causes a class
mismatch when we attempt to re-hydrate a legacy runnable class in AdapterWorkAction.

We now ensure that the context classloader for no-isolation workers get set
to the context classloader of the thread that submitted the work.