Setters for legacy properties backed by Property clear conventions
The version property of a archive task by convention is set to
project.version.

In previous releases, archiveTask.version = null would clear the
value of version and we would strip this from the archive file name.

In Gradle 5.6, setting a Property's value to null reverts back to
the convention's value.

To mimic the old behavior, using the legacy setter now overwrites
both the value and the convention.  So when setting the property
explicitly to null, the underlying Property now has a null value.