Do not drop variant attributes in results based on maven artifacts
FixedComponentArtifacts dropped the variant attributes (stored in
ConfigurationMetadata) for no clear reason. Because of this, the
attributes in the resolve result differed depending on whether the
variant was constructed from pom or GMM.
This is only affecting the attributes reported in the result. During
matching, which happens earlier, all attributes were already considered.