Turn 'multiple publications with same coordinates' error into warning
This reduces the severity of the error detection introduced in #9466.
Although this situation is technically an error that should be avoided
as it causes information loss in the published result, in some testing
scenarios builds may rely on this behavior. So instead of failing,
Gradle now prints a warning message.